download ghg chart table csv refactor
